Tidy up: moved method to Settings and renamed it.
